Fixed some formatting + cleanup
Weapon type is a bit different because it isn't used for image retrieval and is lumped in with CollectableType when the file is read. This means the switch case is the safest method for differentiating the weapons. Cleaned up some of the formatting and got rid of temporary comments. This should function as intended.